moving scenario_result
