+++ pyrit/models/scenario_result.py
